+++ crates/audio/examples/pan.rs
@@ -1,0 +1,47 @@
+use std::time::Duration;
+
+use ambient_audio::{track::Track, AudioStream, Source};
+
+fn main() {
+    let stream = AudioStream::new().unwrap();
+
+    let mixer = stream.mixer();
+
+    // let ambience = Track::from_wav(
+    //     Cursor::new(include_bytes!("../assets/ambience.wav")),
+    //     "Ambience".into(),
+    // )
+    // .unwrap()
+    // .as_source();
+
+    let source = Track::from_vorbis(
+        std::fs::read("../../guest/rust/examples/basics/physics/assets/bonk.ogg")
+            .unwrap()
+            .to_vec(),
+    )
+    .unwrap();
+
+    let source = source
+        .decode()
+        .repeat()
+        .pan(0.0)
+        .take(Duration::from_secs(5))
+        .chain(
+            source
+                .decode()
+                .repeat()
+                .pan(-0.999)
+                .take(Duration::from_secs(5)),
+        )
+        .chain(
+            source
+                .decode()
+                .repeat()
+                .pan(1.0)
+                .take(Duration::from_secs(5)),
+        );
+    // .repeat();
+
+    let sound = mixer.play(source);
+    sound.wait_blocking();
+}