+++ crates/wasm/src/shared/implementation/world_audio.rs
@@ -32,10 +32,14 @@ pub(crate) fn set_listener(world: &mut World, entity: wit::types::EntityId) -> a
     Ok(())
 }
 
-pub(crate) fn set_emitter(world: &mut World, entity: wit::types::EntityId) -> anyhow::Result<()> {
+pub(crate) fn set_emitter(
+    world: &mut World,
+    entity: wit::types::EntityId,
+    volume: f32,
+) -> anyhow::Result<()> {
     let pos = world.get(entity.from_bindgen(), translation())?;
     let emitter = Arc::new(Mutex::new(AudioEmitter {
-        amplitude: 5.0,
+        amplitude: volume,
         attenuation: Attenuation::InversePoly {
             quad: 0.1,
             lin: 0.0,