+++ guest/rust/examples/ui/image/src/client.rs
@@ -20,18 +20,18 @@ fn App(_hooks: &mut Hooks) -> Element {
         .with_default(size_from_background_image()),
         FlowRow::el([
             ImageFromUrl {
-                url: "https://upload.wikimedia.org/wikipedia/commons/b/bf/Bucephala-albeola-010.jpg"
+                url: "https://commons.wikimedia.org/w/index.php?title=Special:Redirect/file/Bucephala-albeola-010.jpg"
                     .to_string(),
             }
             .el(),
             ImageFromUrl {
-                url: "https://upload.wikimedia.org/wikipedia/commons/9/9d/Alpha_transparency_image.png"
+                url: "https://commons.wikimedia.org/w/index.php?title=Special:Redirect/file/Alpha_transparency_image.png"
                     .to_string(),
             }
             .el()
             .with(background_color(), vec4(1., 0., 1., 1.)),
             ImageFromUrl {
-                url: "https://upload.wikimedia.org/wikipedia/commons/b/bf/Bucephala-albeola-010.jpg"
+                url: "https://commons.wikimedia.org/w/index.php?title=Special:Redirect/file/Bucephala-albeola-010.jpg"
                     .to_string(),
             }
             .el()